Reorganize files and translate login to portuguese
