+++ assets/src/components/files/Editor.vue
