+++ assets/src/components/files/ListingItem.vue
