+++ assets/src/i18n/en.json
@@ -1,0 +1,8 @@
+{
+  "login": {
+    "wrongCredentials": "Wrong credentials",
+    "username": "Username",
+    "password": "Password",
+    "submit": "Login"
+  }
+}
