+++ assets/src/main.js
@@ -2,6 +2,7 @@ import Vue from 'vue'
 import App from './App'
 import store from './store'
 import router from './router'
+import i18n from './i18n'
 
 Vue.config.productionTip = true
 
@@ -10,6 +11,7 @@ new Vue({
   el: '#app',
   store,
   router,
+  i18n,
   template: '<App/>',
   components: { App }
 })
