+++ package.json
@@ -14,6 +14,7 @@
     "moment": "^2.18.1",
     "normalize.css": "^7.0.0",
     "vue": "^2.3.3",
+    "vue-i18n": "^7.1.0",
     "vue-router": "^2.7.0",
     "vuex": "^2.3.1"
   },
